move code from Main ScalaBackend with tests
