--- test/Main.hs
@@ -1,4 +1,0 @@
-module Main (main) where
-
-main :: IO ()
-main = putStrLn "Test suite not yet implemented."
